feat(sandbox): extract the host backend as the first plugin (Phase B)
The host backend is now a plugin package against the contract, proving the
Phase A seam end to end:

- packages/sandbox-contract: the contract package — SandboxAdapter (open
  over { type: string } configs), Executor (generic over its config), the
  context/capability types, and the shared createMountedRuntimePathContext
  / killProcessTree helpers. src/sandbox/types.ts re-exports the types with
  the daemon's closed built-in union as the default config, so existing
  daemon import sites are unchanged.
- packages/sandbox-host: HostExecutor + parseHostSandboxArg +
  hostSandboxAdapter (moved from src/sandbox/host.ts), implementing the
  contract; 5 package tests.
- daemon: src/sandbox/index.ts imports and registers the plugin adapter
  from the package; container/firecracker/gondolin now import the shared
  helpers and HostExecutor from the packages; src/sandbox/host.ts is gone.
- embed: sandbox-contract + sandbox-host added to the embed step so the
  published npm daemon stays self-contained (verified: check-npm-package
  passes, 1203 KiB tarball).

Verification: 124 files / 1743 tests, oxlint 0/0, oxfmt clean, knip clean.

A new sandbox backend is now: implement the contract in a package, register
one line in the daemon — no core switch changes.
