+++ packages/sandbox-contract/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@geminixiang/mikan-sandbox-contract",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Sandbox backend plugin contract for mikan: the SandboxAdapter surface, executor types, and shared path/process helpers that both the daemon core and sandbox plugins compile against",
+  "license": "MIT",
+  "files": [
+    "lib"
+  ],
+  "type": "module",
+  "main": "./lib/index.js",
+  "types": "./lib/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./lib/index.d.ts",
+      "default": "./lib/index.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "tsgo -p tsconfig.build.json"
+  },
+  "devDependencies": {
+    "@types/node": "^26.0.1",
+    "@typescript/native-preview": "^7.0.0-dev.20260624.1",
+    "typescript": "^6.0.3"
+  }
+}
