+++ src/sandbox/gondolin.ts
@@ -14,13 +14,8 @@ import * as log from "../log.js";
 import type { ResourceLimits, SandboxLimitStatus, SandboxResourceController } from "../types.js";
 import { withRuntimeBootstrap } from "./container.js";
 import { SandboxError } from "./errors.js";
-import {
-  createMountedRuntimePathContext,
-  execReadFile,
-  execReadFileBase64,
-  execWriteFile,
-  shellEscape,
-} from "./utils.js";
+import { createMountedRuntimePathContext } from "@geminixiang/mikan-sandbox-contract";
+import { execReadFile, execReadFileBase64, execWriteFile, shellEscape } from "./utils.js";
 import type * as GondolinNamespace from "@earendil-works/gondolin";
 import type {
   ExecOptions,
