+++ src/sandbox/utils.ts
@@ -1,6 +1,4 @@
 import { spawn } from "node:child_process";
-import { isAbsolute, relative, resolve as resolvePath, sep } from "node:path";
-import type { RuntimePathContext } from "./types.js";
 
 export function execSimple(cmd: string, args: string[]): Promise<string> {
   return new Promise((resolve, reject) => {
@@ -18,29 +16,6 @@ export function execSimple(cmd: string, args: string[]): Promise<string> {
       else reject(new Error(stderr || `Exit code ${code}`));
     });
   });
-}
-
-export function killProcessTree(pid: number): void {
-  if (process.platform === "win32") {
-    try {
-      spawn("taskkill", ["/F", "/T", "/PID", String(pid)], {
-        stdio: "ignore",
-        detached: true,
-      });
-    } catch {
-      // Ignore errors
-    }
-  } else {
-    try {
-      process.kill(-pid, "SIGKILL");
-    } catch {
-      try {
-        process.kill(pid, "SIGKILL");
-      } catch {
-        // Process already dead
-      }
-    }
-  }
 }
 
 export function shellEscape(s: string): string {
@@ -115,48 +90,4 @@ export async function execWriteFile<TOptions>(
   await run(
     `base64 -d < ${stageB64} > ${stage} && mv ${stage} ${escapedPath} && rm -f ${stageB64}`,
   );
-}
-
-export function createMountedRuntimePathContext(
-  hostWorkspaceRoot: string,
-  runtimeWorkspaceRoot: string,
-): RuntimePathContext {
-  return {
-    hostWorkspaceRoot,
-    runtimeWorkspaceRoot,
-    runtimeToHostPath: (runtimePath) =>
-      translateMountedRuntimePathToHost(runtimePath, runtimeWorkspaceRoot, hostWorkspaceRoot),
-  };
-}
-
-function translateMountedRuntimePathToHost(
-  runtimePath: string,
-  runtimeWorkspaceRoot: string,
-  hostWorkspaceRoot: string,
-): string {
-  if (!isAbsolute(runtimePath)) {
-    return runtimePath;
-  }
-
-  const runtimeRoot = resolvePath(runtimeWorkspaceRoot);
-  const normalizedRuntimePath = resolvePath(runtimePath);
-  const runtimeRelativePath = relative(runtimeRoot, normalizedRuntimePath);
-  const escapesRuntimeRoot =
-    runtimeRelativePath === ".." ||
-    runtimeRelativePath.startsWith(`..${sep}`) ||
-    isAbsolute(runtimeRelativePath);
-
-  if (escapesRuntimeRoot) {
-    return runtimePath;
-  }
-
-  const hostRoot = resolvePath(hostWorkspaceRoot);
-  const hostPath = resolvePath(hostRoot, runtimeRelativePath);
-  const hostRelativePath = relative(hostRoot, hostPath);
-  const escapesHostRoot =
-    hostRelativePath === ".." ||
-    hostRelativePath.startsWith(`..${sep}`) ||
-    isAbsolute(hostRelativePath);
-
-  return escapesHostRoot ? runtimePath : hostPath;
 }
